Conflicts:
	CBP/android/sub_app/fermat-cbp-android-sub-app-crypto-broker-community-bitdubai/src/main/java/com/bitdubai/sub_app/crypto_broker_community/common/holders/AvailableActorsViewHolder.java
	CBP/plugin/actor_network_service/fermat-cbp-plugin-actor-network-service-crypto-broker-bitdubai/src/main/java/com/bitdubai/fermat_cbp_plugin/layer/actor_network_service/crypto_broker/developer/bitdubai/version_1/structure/CryptoBrokerActorNetworkServiceSearch.java

+++ CER/plugin/provider/fermat-cer-plugin-provider-dolartoday-bitdubai/src/main/java/com/bitdubai/fermat_cer_plugin/layer/provider/dolartoday/developer/bitdubai/version_1/ProviderDolarTodayPluginRoot.java
@@ -132,7 +132,7 @@ public class ProviderDolarTodayPluginRoot extends AbstractPlugin implements Data
         if(currencyPair.getFrom() == FiatCurrency.US_DOLLAR)
             purchasePrice = salePrice = 1150;
         else
-            purchasePrice = salePrice = 1/1150;
+            purchasePrice = salePrice = 1/ (double) 1150;
 
 //        double purchasePrice = 0;
 //        double salePrice = 0;
